Move BookingDetailsCard to BooknowCards directory.
